services/document-service/requirements.txt
fastapi
uvicorn[standard]
python-multipart
psycopg2-binary
python-dotenv
bcrypt
python-jose[cryptography]
